OIDN/requirements.txt
absl-py==1.4.0
accelerate==0.16.0
asgiref==3.7.2
attrs==23.2.0
av==10.0.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
cachetools==5.3.1
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.0.1
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython==3.0.10
debugpy @ file:///tmp/build/80754af9/debugpy_1637091426235/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
Django==3.2.23
einops==0.6.0
ema-pytorch==0.2.1
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649908585034/work
exceptiongroup==1.2.1
filelock==3.12.2
fonttools==4.25.0
fsspec==2023.1.0
future==1.0.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
grpcio==1.56.0
huggingface-hub==0.16.4
hypothesis==6.79.4
idna==3.4
image==1.5.33
imageio==2.31.1
importlib-metadata==6.0.0
ipykernel @ file:///opt/conda/conda-bld/ipykernel_1662361798230/work
ipython @ file:///tmp/abs_94gruux8u8/croots/recipe/ipython_1659529858706/work
jedi @ file:///tmp/build/80754af9/jedi_1644299024593/work
jupyter_client @ file:///croot/jupyter_client_1676329080601/work
jupyter_core @ file:///croot/jupyter_core_1668084443243/work
kiwisolver @ file:///opt/conda/conda-bld/kiwisolver_1638569886207/work
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
munkres==1.1.4
nest-asyncio @ file:///croot/nest-asyncio_1672387112409/work
networkx==2.6.3
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-contrib-python==4.0.0.21
opencv-python==4.0.0.21
packaging @ file:///croot/packaging_1671697413597/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.4.0
pimm==0.0.6
ping3==4.0.8
ply==3.11
prompt-toolkit @ file:///croot/prompt-toolkit_1672387306916/work
protobuf==3.20.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptflops==0.7.3
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5-sip==12.11.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytorch-msssim==1.0.0
pytz==2023.3.post1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq @ file:///opt/conda/conda-bld/pyzmq_1657724186960/work
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.3
scikit-image==0.16.2
scipy==1.7.3
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sortedcontainers==2.4.0
sqlparse==0.4.4
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
tifffile==2021.11.2
timm==0.9.12
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
torch==1.9.1+cu111
torchaudio==0.9.1
torchpwl==0.2.0
torchvision==0.10.1+cu111
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm==4.64.1
traitlets @ file:///croot/traitlets_1671143879854/work
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug==2.2.3
zipp==3.14.0
